Move ChessBotVictor into separate file
